+++ src/components/highlight/__tests__/highlight-test-cases.js
@@ -10,126 +10,10 @@ import { highlightXml } from '../xml';
 import { highlightKotlin } from '../kotlin';
 import { highlightCss } from '../css';
 import { highlightGroovy } from '../groovy';
+import { highlightThemeCss } from '../../highlight/theme-css.js';
 
 const testCases = {};
 const noRenderCases = {};
-
-const css = `code[class*='language-'],
-pre[class*='language-'] {
-  color: #273d56;
-  background: none;
-  font-family: Consolas, Monaco, 'Andale Mono', 'Ubuntu Mono', monospace;
-  text-align: left;
-  white-space: pre;
-  word-spacing: normal;
-  word-break: normal;
-  word-wrap: normal;
-  line-height: 1.5;
-
-  -moz-tab-size: 4;
-  -o-tab-size: 4;
-  tab-size: 4;
-
-  -webkit-hyphens: none;
-  -moz-hyphens: none;
-  -ms-hyphens: none;
-  hyphens: none;
-}
-
-/* Code blocks */
-pre[class*='language-'] {
-  padding: 1em;
-  margin: 0.5em 0;
-  overflow: auto;
-  border-radius: 0.3em;
-}
-
-:not(pre) > code[class*='language-'],
-pre[class*='language-'] {
-  background: #272822;
-}
-
-/* Inline code */
-:not(pre) > code[class*='language-'] {
-  padding: 0.1em;
-  border-radius: 0.3em;
-  white-space: normal;
-}
-
-.token.comment,
-.token.prolog,
-.token.doctype,
-.token.cdata {
-  color: #53708e;
-}
-
-.token.punctuation {
-  color: #273d56;
-}
-
-.namespace {
-  opacity: 0.7;
-}
-
-.token.property,
-.token.tag,
-.token.constant,
-.token.symbol,
-.token.deleted {
-  color: #314ccd;
-}
-
-.token.boolean,
-.token.number {
-  color: #7753eb;
-}
-
-.token.selector,
-.token.attr-name,
-.token.string,
-.token.char,
-.token.builtin,
-.token.inserted {
-  color: #ce2c69;
-}
-
-.token.operator,
-.token.entity,
-.token.url,
-.language-css .token.string,
-.style .token.string,
-.token.variable {
-  color: #273d56;
-}
-
-.token.atrule,
-.token.attr-value,
-.token.function,
-.token.class-name {
-  color: #4264fb;
-}
-
-.token.keyword {
-  color: #314ccd;
-}
-
-.token.regex,
-.token.important {
-  color: #fd971f;
-}
-
-.token.important,
-.token.bold {
-  font-weight: bold;
-}
-.token.italic {
-  font-style: italic;
-}
-
-.token.entity {
-  cursor: help;
-}
-`;
 
 const swiftCodeSnippet = `
 import Mapbox
@@ -310,7 +194,7 @@ testCases.basic = {
   props: {
     code: swiftCodeSnippet,
     highlightedCode: highlightSwift(swiftCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -320,7 +204,7 @@ testCases.objc = {
   props: {
     code: objectiveCCodeSnippet,
     highlightedCode: highlightObjectivec(objectiveCCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -330,7 +214,7 @@ testCases.java = {
   props: {
     code: javaCodeSnippet,
     highlightedCode: highlightJava(javaCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -340,7 +224,7 @@ testCases.json = {
   props: {
     code: jsonCodeSnippet,
     highlightedCode: highlightJson(jsonCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -350,7 +234,7 @@ testCases.jsx = {
   props: {
     code: jsxCodeSnippet,
     highlightedCode: highlightJsx(jsxCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -360,7 +244,7 @@ testCases.html = {
   props: {
     code: htmlCodeSnippet,
     highlightedCode: highlightHtml(htmlCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -370,7 +254,7 @@ testCases.xml = {
   props: {
     code: xmlCodeSnippet,
     highlightedCode: highlightXml(xmlCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -380,7 +264,7 @@ testCases.kotlin = {
   props: {
     code: kotlinCodeSnippet,
     highlightedCode: highlightKotlin(kotlinCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -390,7 +274,7 @@ testCases.css = {
   props: {
     code: cssCodeSnippet,
     highlightedCode: highlightCss(cssCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };
 
@@ -400,7 +284,7 @@ testCases.groovy = {
   props: {
     code: groovyCodeSnippet,
     highlightedCode: highlightGroovy(groovyCodeSnippet),
-    highlightThemeCss: css
+    highlightThemeCss: highlightThemeCss
   }
 };